Separated SpecFlow/NUnit and MSUnit tests in ACKTests project, created test class for DirectParse (parser level).
