+++ AKCTests/SpecFlow/ClippingTests/Clippings.cs
