+++ AKCTests/SpecFlow/ClippingTests/Parsing_EntrySteps.cs
