+++ AKCWPF/GUI/MainWindow.xaml.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using AKCCore;
+using System.Linq;
 
 namespace AKCWPF {
     /// <summary>
@@ -120,6 +121,16 @@ namespace AKCWPF {
                 }
 
                 if (correctParserConfirmed) {
+                    /*
+                    //TEST
+                    var test = parserController.options.SelectedParser.DirectParse(
+                        "A Dance With Dragons: Book 5 of A Song of Ice and Fire(Song of Ice & Fire 5)(Martin, George R.R.)\r\n- Mi subrayado en la página 41 | Posición 620 - 620 | Añadido el domingo 2 de septiembre de 2012, 23:54:20\r\n \r\n droll?”\r\n==========\r\nLa metamorfosis (Spanish Edition) (Kafka, Franz)\r\n- Mi subrayado Posición 13 - 13 | Añadido el jueves 18 de octubre de 2012, 19:18:28\r\ndespués\r\nhabía\r\nmonos\r\n \r\n==========\r\nA Dance With Dragons: Book 5 of A Song of Ice and Fire(Song of Ice & Fire 5)(Martin, George R.R.)\r\n- Mi marcador en la página 442 | Posición 6769 | Añadido el miércoles 28 de noviembre de 2012, 14:08:48\r\n \r\n==========",
+                         parserController.options.SelectedFormat
+                        );
+
+                    var myResult = test.ToList();
+                    */
+
                     //Async parsing
                     LW = new LoadingWindow();
                     await Task.Run(() => parserController.RunParser(parserController.options.TextToParsePath));
